+++ packages/i18n/@types/env.d.ts
@@ -1,5 +1,0 @@
-import type { LocaleConfig } from '../src'
-
-declare global {
-  var I18N: Record<string, LocaleConfig>
-}