+++ packages/web/src/__scripts__/postinstall.ts
@@ -40,7 +40,7 @@ declare module '@vue/runtime-core' {
     t: TemplateFunctions['t']
     viewTitle: string
     viewIcon: string
-    instanceConfig: typeof import('aeria-ui-build').InstanceConfig
+    instanceVars: typeof import('aeria-ui-build').InstanceConfig['site']
     currentUser: (Collections['user']['item'] extends infer UserCollection
       ? UserCollection extends (...args: any[]) => any
         ? ReturnType<UserCollection>