+++ packages/web/src/options.ts
@@ -1,7 +1,7 @@
 import type { App, Component } from 'vue'
 import type { RouteRecordRaw } from 'vue-router'
 import type { I18nConfig } from '@aeria-ui/i18n'
-import type { GlobalStateManager } from '@aeria-ui/state-management'
+import type { StoreContext } from '@aeria-ui/state-management'
 import type { RouteMeta } from './router.js'
 
 export type MenuNodeBase = Partial<RouteMeta> & {
@@ -39,7 +39,9 @@ export type AppOptions = {
   routes?: RouteRecordRaw[]
   setup?: (config: {
     app: App
-    globalStateManager: GlobalStateManager
+    context: StoreContext<{
+      i18n: I18nConfig
+    }>
   })=> void | Promise<void>
 }