+++ packages/web/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @aeria-ui/web
 
+## 0.0.18
+
+### Patch Changes
+
+- Use p/i in i18n
+- Updated dependencies
+  - @aeria-ui/state-management@0.0.7
+  - @aeria-ui/i18n@0.0.1
+
 ## 0.0.17
 
 ### Patch Changes